Extract CmuxMobilePairedMac + CmuxMobileRPC from cmuxFeature (iOS refactor 5079 wave 1)
Move the SQLite paired-mac actor into Packages/CmuxMobilePairedMac behind a
MobilePairedMacStoring seam, drop MobileShellStorePairedMacStoreFactory, and
open/inject the store at CMUXMobileRootScene. Move MobileCoreRPCClient/Session +
ticket decode + DTOs + the MobileSyncing seam into Packages/CmuxMobileRPC, with
a MobileSyncRuntime protocol so RPC depends only on the wire core (CMUXMobileRuntime
conforms). Relocate the RPC-only tests + transport doubles into the new test target.
