+++ Packages/CmuxMobileRPC/Package.swift
@@ -1,0 +1,50 @@
+// swift-tools-version: 6.0
+
+import PackageDescription
+
+let package = Package(
+    name: "CmuxMobileRPC",
+    platforms: [
+        .iOS(.v18),
+        .macOS(.v14),
+    ],
+    products: [
+        .library(
+            name: "CmuxMobileRPC",
+            targets: ["CmuxMobileRPC"]
+        ),
+    ],
+    dependencies: [
+        .package(path: "../CMUXMobileCore"),
+        .package(path: "../CmuxMobileShellModel"),
+        .package(path: "../CmuxMobileSupport"),
+    ],
+    targets: [
+        .target(
+            name: "CmuxMobileRPC",
+            dependencies: [
+                "CMUXMobileCore",
+                "CmuxMobileShellModel",
+                "CmuxMobileSupport",
+            ],
+            swiftSettings: [
+                .swiftLanguageMode(.v6),
+                .enableUpcomingFeature("ExistentialAny"),
+                .enableUpcomingFeature("InternalImportsByDefault"),
+            ]
+        ),
+        .testTarget(
+            name: "CmuxMobileRPCTests",
+            dependencies: [
+                "CmuxMobileRPC",
+                "CMUXMobileCore",
+                "CmuxMobileShellModel",
+            ],
+            swiftSettings: [
+                .swiftLanguageMode(.v6),
+                .enableUpcomingFeature("ExistentialAny"),
+                .enableUpcomingFeature("InternalImportsByDefault"),
+            ]
+        ),
+    ]
+)
