+++ Packages/CmuxMobileRPC/Sources/CmuxMobileRPC/MobileCoreRPCClient.swift
@@ -1,0 +1,278 @@
+public import CMUXMobileCore
+internal import CmuxMobileShellModel
+internal import CmuxMobileSupport
+public import Foundation
+
+/// A multiplexed RPC client over a single persistent transport to a paired Mac.
+///
+/// All stored properties are immutable `let`s of `Sendable` types (the session
+/// is an actor), so this is genuinely `Sendable` without opting out of checking.
+public final class MobileCoreRPCClient: MobileSyncing, Sendable {
+    private let runtime: any MobileSyncRuntime
+    private let route: CmxAttachRoute
+    private let ticket: CmxAttachTicket
+    private let allowsStackAuthFallback: Bool
+    private let session: MobileCoreRPCSession
+
+    /// Create a client bound to one route + attach ticket.
+    /// - Parameters:
+    ///   - runtime: The DI runtime supplying transport factory, token provider, timeouts, clock.
+    ///   - route: The attach route this client connects over.
+    ///   - ticket: The attach ticket authorizing requests.
+    ///   - allowsStackAuthFallback: When `true`, falls back to a Stack Auth token
+    ///     on routes that allow it once the attach ticket no longer covers a request.
+    public init(
+        runtime: any MobileSyncRuntime,
+        route: CmxAttachRoute,
+        ticket: CmxAttachTicket,
+        allowsStackAuthFallback: Bool = false
+    ) {
+        self.runtime = runtime
+        self.route = route
+        self.ticket = ticket
+        self.allowsStackAuthFallback = allowsStackAuthFallback
+        self.session = MobileCoreRPCSession(
+            makeTransport: { [route, runtime] in
+                try runtime.transportFactory.makeTransport(for: route)
+            }
+        )
+    }
+
+    /// Tear down the persistent transport (called when the client is
+    /// replaced or the user signs out).
+    public func disconnect() async {
+        await session.tearDown(error: .connectionClosed)
+    }
+
+    /// Subscribe to server-pushed events. Returns a stream of envelopes
+    /// matching any of the requested topics. Cancel by terminating iteration.
+    public func subscribe(to topics: Set<String>) async -> AsyncStream<MobileEventEnvelope> {
+        await session.addEventListener(topics: topics).stream
+    }
+
+    /// Build a JSON-RPC request frame with the given method and params.
+    /// - Parameters:
+    ///   - method: The RPC method name.
+    ///   - params: The request parameters.
+    ///   - id: The request id (defaults to a fresh UUID).
+    /// - Returns: The encoded request data.
+    /// - Throws: A serialization error if the params are not JSON-encodable.
+    public static func requestData(
+        method: String,
+        params: [String: Any] = [:],
+        id: String = UUID().uuidString
+    ) throws -> Data {
+        let request: [String: Any] = [
+            "id": id,
+            "method": method,
+            "params": params,
+        ]
+        return try JSONSerialization.data(withJSONObject: request)
+    }
+
+    public func sendRequest(_ requestData: Data, timeoutNanoseconds: UInt64? = nil) async throws -> Data {
+        // Multiplexed over a persistent transport: each request gets a unique
+        // id, the session's reader task routes the response back here. No
+        // connect/close per RPC, no head-of-line blocking between calls.
+        let (id, augmented) = try Self.requestWithGuaranteedID(requestData)
+        let authenticated = try await requestDataWithAuth(augmented)
+        return try await Self.withRequestTimeout(
+            timeoutNanoseconds: timeoutNanoseconds ?? runtime.rpcRequestTimeoutNanoseconds
+        ) {
+            try await self.session.send(payload: authenticated, requestID: id)
+        }
+    }
+
+    private static func requestWithGuaranteedID(_ requestData: Data) throws -> (String, Data) {
+        guard var dict = try JSONSerialization.jsonObject(with: requestData) as? [String: Any] else {
+            throw MobileShellConnectionError.invalidResponse
+        }
+        let id: String
+        if let existing = dict["id"] as? String, !existing.isEmpty {
+            id = existing
+        } else {
+            id = UUID().uuidString
+            dict["id"] = id
+        }
+        let data = try JSONSerialization.data(withJSONObject: dict)
+        return (id, data)
+    }
+
+    private func requestDataWithAuth(_ requestData: Data) async throws -> Data {
+        guard var request = try JSONSerialization.jsonObject(with: requestData) as? [String: Any] else {
+            return requestData
+        }
+        let requestNeedsAuth = Self.requestRequiresAuth(request)
+        let requestIsCoveredByAttachTicket = !Self.requestNeedsStackAuthFallback(request, ticket: ticket)
+        var auth: [String: Any] = [:]
+        let attachToken = ticket.authToken?.trimmingCharacters(in: .whitespacesAndNewlines)
+        let hasAttachToken = attachToken?.isEmpty == false
+        if let attachToken,
+           requestNeedsAuth,
+           hasAttachToken,
+           requestIsCoveredByAttachTicket {
+            if ticket.expiresAt > runtime.now() {
+                auth["attach_token"] = attachToken
+            } else if !allowsStackAuthFallback || !MobileShellRouteAuthPolicy.routeAllowsStackAuth(route) {
+                throw MobileShellConnectionError.attachTicketExpired
+            }
+        }
+        let shouldSendStackAuth = requestNeedsAuth && auth["attach_token"] == nil
+        if shouldSendStackAuth {
+            guard allowsStackAuthFallback,
+                  MobileShellRouteAuthPolicy.routeAllowsStackAuth(route) else {
+                throw MobileShellConnectionError.insecureManualRoute
+            }
+            do {
+                auth["stack_access_token"] = try await runtime.stackAccessTokenProvider()
+            } catch {
+                throw MobileShellConnectionError.authorizationFailed(
+                    L10n.string(
+                        "mobile.pairing.stackAuthTokenUnavailable",
+                        defaultValue: "Sign in on your computer with the same account, then try again."
+                    )
+                )
+            }
+        }
+        if !auth.isEmpty {
+            request["auth"] = auth
+        }
+        return try JSONSerialization.data(withJSONObject: request)
+    }
+
+    private static func requestNeedsStackAuthFallback(_ request: [String: Any], ticket: CmxAttachTicket) -> Bool {
+        guard requestRequiresAuth(request) else {
+            return false
+        }
+        let method = (request["method"] as? String)?.trimmingCharacters(in: .whitespacesAndNewlines)
+        let params = request["params"] as? [String: Any] ?? [:]
+        let workspaceSelection = stringParamSelection(params, keys: ["workspace_id"])
+        let terminalSelection = stringParamSelection(params, keys: ["surface_id", "terminal_id", "tab_id"])
+        if workspaceSelection.hasConflict ||
+            terminalSelection.hasConflict ||
+            containsIgnoredAliasParameters(params) {
+            return true
+        }
+
+        switch method {
+        case "mobile.workspace.list", "workspace.list":
+            return false
+        case "workspace.create":
+            return false
+        case "mobile.terminal.create", "terminal.create":
+            return false
+        case "mobile.terminal.input", "terminal.input",
+             "mobile.terminal.replay", "terminal.replay",
+             "mobile.terminal.viewport", "terminal.viewport":
+            return !ticketCoversTerminalRequest(
+                ticket: ticket,
+                workspaceSelection: workspaceSelection.value,
+                terminalSelection: terminalSelection.value
+            )
+        case "mobile.events.subscribe", "mobile.events.unsubscribe":
+            return false
+        default:
+            return true
+        }
+    }
+
+    private static func requestRequiresAuth(_ request: [String: Any]) -> Bool {
+        let method = (request["method"] as? String)?.trimmingCharacters(in: .whitespacesAndNewlines)
+        // Only the unauthenticated host probe is exempt. attach_ticket.create has no
+        // attach token yet (it mints the ticket), so requiring auth routes it through
+        // the Stack Auth account token: a ticket can only be created by a signed-in user.
+        return method != "mobile.host.status"
+    }
+
+    private static func ticketCoversTerminalRequest(
+        ticket: CmxAttachTicket,
+        workspaceSelection: String?,
+        terminalSelection: String?
+    ) -> Bool {
+        let ticketWorkspaceID = ticket.workspaceID.trimmingCharacters(in: .whitespacesAndNewlines)
+        // Empty workspaceID means the ticket is Mac-wide (general pairing).
+        // It covers any workspace/terminal on the paired Mac.
+        if ticketWorkspaceID.isEmpty {
+            return true
+        }
+        if let workspaceSelection, workspaceSelection != ticketWorkspaceID {
+            return false
+        }
+
+        if let ticketTerminalID = ticket.terminalID?.trimmingCharacters(in: .whitespacesAndNewlines),
+           !ticketTerminalID.isEmpty {
+            return terminalSelection == ticketTerminalID
+        }
+
+        return workspaceSelection == ticketWorkspaceID
+    }
+
+    private static func containsIgnoredAliasParameters(_ params: [String: Any]) -> Bool {
+        params["workspaceID"] != nil || params["terminalID"] != nil
+    }
+
+    private static func stringParamSelection(
+        _ params: [String: Any],
+        keys: [String]
+    ) -> StringParamSelection {
+        var selected: String?
+        for key in keys {
+            if let value = params[key] as? String {
+                let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
+                if !trimmed.isEmpty {
+                    if let selected, selected != trimmed {
+                        return StringParamSelection(value: selected, hasConflict: true)
+                    }
+                    selected = selected ?? trimmed
+                }
+            }
+        }
+        return StringParamSelection(value: selected, hasConflict: false)
+    }
+
+    private struct StringParamSelection {
+        let value: String?
+        let hasConflict: Bool
+    }
+
+    private static func withRequestTimeout<T: Sendable>(
+        timeoutNanoseconds: UInt64,
+        operation: @escaping @Sendable () async throws -> T
+    ) async throws -> T {
+        try Task.checkCancellation()
+        return try await withThrowingTaskGroup(of: T.self) { group in
+            group.addTask {
+                try await operation()
+            }
+            group.addTask {
+                try await Task.sleep(nanoseconds: timeoutNanoseconds)
+                throw MobileShellConnectionError.requestTimedOut
+            }
+            do {
+                guard let result = try await group.next() else {
+                    throw CancellationError()
+                }
+                group.cancelAll()
+                return result
+            } catch {
+                group.cancelAll()
+                throw error
+            }
+        }
+    }
+}
+
+#if DEBUG
+extension MobileCoreRPCClient {
+    /// Test-only hook exposing the private request-timeout race for unit tests.
+    public static func debugWithRequestTimeout<T: Sendable>(
+        timeoutNanoseconds: UInt64,
+        operation: @escaping @Sendable () async throws -> T
+    ) async throws -> T {
+        try await withRequestTimeout(
+            timeoutNanoseconds: timeoutNanoseconds,
+            operation: operation
+        )
+    }
+}
+#endif
