+++ Packages/CmuxMobileRPC/Tests/CmuxMobileRPCTests/TransportTestDoubles.swift
@@ -1,0 +1,157 @@
+import CMUXMobileCore
+import Foundation
+import Testing
+@testable import CmuxMobileRPC
+
+/// A minimal `MobileSyncRuntime` for tests, supplying a transport factory,
+/// stack-token provider, timeout, and clock without pulling the app's DI bundle.
+struct TestMobileSyncRuntime: MobileSyncRuntime {
+    var transportFactory: any CmxByteTransportFactory
+    var stackAccessTokenProvider: @Sendable () async throws -> String
+    var rpcRequestTimeoutNanoseconds: UInt64
+    var now: @Sendable () -> Date
+
+    init(
+        transportFactory: any CmxByteTransportFactory,
+        stackAccessToken: String? = "test-stack-token",
+        rpcRequestTimeoutNanoseconds: UInt64 = 30 * 1_000_000_000,
+        now: @escaping @Sendable () -> Date = Date.init
+    ) {
+        self.transportFactory = transportFactory
+        self.stackAccessTokenProvider = {
+            guard let stackAccessToken else { throw MissingTestStackAccessToken() }
+            return stackAccessToken
+        }
+        self.rpcRequestTimeoutNanoseconds = rpcRequestTimeoutNanoseconds
+        self.now = now
+    }
+}
+
+struct MissingTestStackAccessToken: Error {}
+
+/// Async-safe one-shot boolean flag used to observe task progress in tests.
+actor AsyncFlag {
+    private var value = false
+
+    func set() {
+        value = true
+    }
+
+    func isSet() -> Bool {
+        value
+    }
+}
+
+/// A parsed snapshot of one RPC request frame for test assertions.
+struct RecordedRPCRequest: Sendable {
+    var id: String?
+    var method: String?
+    var workspaceID: String?
+    var terminalID: String?
+    var text: String?
+    var hasAuth: Bool
+    var attachToken: String?
+    var stackAccessToken: String?
+}
+
+func recordedRPCRequest(from payload: Data) throws -> RecordedRPCRequest {
+    let request = try #require(JSONSerialization.jsonObject(with: payload) as? [String: Any])
+    let params = request["params"] as? [String: Any] ?? [:]
+    let auth = request["auth"] as? [String: Any]
+    return RecordedRPCRequest(
+        id: request["id"] as? String,
+        method: request["method"] as? String,
+        workspaceID: params["workspace_id"] as? String,
+        terminalID: params["terminal_id"] as? String ?? params["surface_id"] as? String,
+        text: params["text"] as? String,
+        hasAuth: auth != nil,
+        attachToken: auth?["attach_token"] as? String,
+        stackAccessToken: auth?["stack_access_token"] as? String
+    )
+}
+
+func hostPortRoute(
+    kind: CmxAttachTransportKind,
+    host: String,
+    port: Int,
+    priority: Int = 0
+) throws -> CmxAttachRoute {
+    try CmxAttachRoute(
+        id: kind.rawValue,
+        kind: kind,
+        endpoint: .hostPort(host: host, port: port),
+        priority: priority
+    )
+}
+
+/// Transport that blocks its first `send` until released, recording payloads so
+/// tests can assert a cancelled queued request is never written.
+actor QueuedCancellationProbeTransport: CmxByteTransport {
+    private var sentPayloads: [Data] = []
+    private var receiveWaiters: [CheckedContinuation<Data?, Never>] = []
+    private var firstSendRelease: CheckedContinuation<Void, Never>?
+    private var shouldBlockFirstSend = true
+    private var isClosed = false
+
+    func connect() async throws {}
+
+    func receive() async throws -> Data? {
+        if isClosed {
+            return nil
+        }
+        return await withCheckedContinuation { continuation in
+            receiveWaiters.append(continuation)
+        }
+    }
+
+    func send(_ data: Data) async throws {
+        var buffer = data
+        let payloads = try MobileSyncFrameCodec.decodeFrames(from: &buffer)
+        sentPayloads.append(contentsOf: payloads)
+        if shouldBlockFirstSend {
+            shouldBlockFirstSend = false
+            await withCheckedContinuation { continuation in
+                firstSendRelease = continuation
+            }
+        }
+    }
+
+    func close() async {
+        isClosed = true
+        let waiters = receiveWaiters
+        receiveWaiters = []
+        for waiter in waiters {
+            waiter.resume(returning: nil)
+        }
+        releaseFirstSend()
+    }
+
+    func releaseFirstSend() {
+        firstSendRelease?.resume()
+        firstSendRelease = nil
+    }
+
+    func sentRequests() throws -> [RecordedRPCRequest] {
+        try sentPayloads.map(recordedRPCRequest(from:))
+    }
+
+    func waitForSentRequestCount(_ count: Int) async throws -> [RecordedRPCRequest] {
+        var requests: [RecordedRPCRequest] = []
+        for _ in 0..<200 {
+            requests = try sentRequests()
+            if requests.count >= count {
+                return requests
+            }
+            try await Task.sleep(nanoseconds: 1_000_000)
+        }
+        return requests
+    }
+}
+
+struct QueuedCancellationProbeTransportFactory: CmxByteTransportFactory {
+    let transport: QueuedCancellationProbeTransport
+
+    func makeTransport(for route: CmxAttachRoute) throws -> any CmxByteTransport {
+        transport
+    }
+}
