+++ ios/cmuxPackage/Package.swift
@@ -19,6 +19,8 @@ let package = Package(
         .package(path: "../../Packages/CMUXMobileCore"),
         .package(path: "../../Packages/CmuxMobileAuth"),
         .package(path: "../../Packages/CmuxMobileDiagnostics"),
+        .package(path: "../../Packages/CmuxMobilePairedMac"),
+        .package(path: "../../Packages/CmuxMobileRPC"),
         .package(path: "../../Packages/CmuxMobileShellModel"),
         .package(path: "../../Packages/CmuxMobileSupport"),
         .package(path: "../../Packages/CmuxMobileTerminal"),
@@ -34,6 +36,8 @@ let package = Package(
                 "CMUXMobileCore",
                 "CmuxMobileAuth",
                 "CmuxMobileDiagnostics",
+                "CmuxMobilePairedMac",
+                "CmuxMobileRPC",
                 "CmuxMobileShellModel",
                 "CmuxMobileSupport",
                 "CmuxMobileTerminal",
@@ -54,6 +58,8 @@ let package = Package(
                 "CMUXMobileCore",
                 "CmuxMobileAuth",
                 "CmuxMobileDiagnostics",
+                "CmuxMobilePairedMac",
+                "CmuxMobileRPC",
                 "CmuxMobileShellModel",
                 "CmuxMobileSupport",
                 "CmuxMobileTerminal",
