+++ ios/cmuxPackage/Sources/cmuxFeature/CMUXMobileRootScene.swift
@@ -1,9 +1,13 @@
+import CmuxMobilePairedMac
+import Foundation
+import OSLog
 import SwiftUI
 
 #if canImport(UIKit) && DEBUG
 import CmuxMobileTerminal
-import Foundation
 #endif
+
+private let mobileRootSceneLog = Logger(subsystem: "dev.cmux.ios", category: "mobile-root-scene")
 
 /// Top-level mobile scene root.
 ///
@@ -19,11 +23,25 @@ import Foundation
 /// `CMUXMobileCore`; the terminal package stays an implementation detail.
 public struct CMUXMobileRootScene: View {
     private let runtime: CMUXMobileRuntime
+    private let pairedMacStore: (any MobilePairedMacStoring)?
 
     /// Creates the root scene.
     /// - Parameter runtime: The mobile runtime that backs the shell store.
     public init(runtime: CMUXMobileRuntime) {
         self.runtime = runtime
+        // TRANSITIONAL (iOS refactor): open the SQLite paired-mac store at the
+        // composition root and inject it as `any MobilePairedMacStoring`,
+        // replacing the deleted `MobileShellStorePairedMacStoreFactory`
+        // singleton. Opening can fail in a read-only sandbox (tests/previews);
+        // the store degrades to in-memory operation when `nil`.
+        do {
+            self.pairedMacStore = try MobilePairedMacStore()
+        } catch {
+            mobileRootSceneLog.error(
+                "failed to open paired mac store: \(String(describing: error), privacy: .public)"
+            )
+            self.pairedMacStore = nil
+        }
     }
 
     public var body: some View {
@@ -31,10 +49,14 @@ public struct CMUXMobileRootScene: View {
         if ProcessInfo.processInfo.environment["CMUX_ZOOM_STRESS"] == "1" {
             MobileZoomStressView()
         } else {
-            CMUXMobileAppView(store: CMUXMobileShellStore(runtime: runtime))
+            CMUXMobileAppView(store: makeStore())
         }
         #else
-        CMUXMobileAppView(store: CMUXMobileShellStore(runtime: runtime))
+        CMUXMobileAppView(store: makeStore())
         #endif
+    }
+
+    private func makeStore() -> CMUXMobileShellStore {
+        CMUXMobileShellStore(runtime: runtime, pairedMacStore: pairedMacStore)
     }
 }
