--- ios/cmuxPackage/Sources/cmuxFeature/MobileCoreRPC.swift
@@ -1,764 +1,0 @@
-import CMUXMobileCore
-import CmuxMobileShellModel
-import CmuxMobileSupport
-import Foundation
-
-enum MobileShellConnectionError: LocalizedError {
-    case invalidResponse
-    case connectionClosed
-    case requestTimedOut
-    case insecureManualRoute
-    case attachTicketExpired
-    case authorizationFailed(String)
-    case rpcError(String?, String)
-
-    var errorDescription: String? {
-        switch self {
-        case .invalidResponse:
-            return "Invalid mobile sync response"
-        case .connectionClosed:
-            return "Mobile sync connection closed"
-        case .requestTimedOut:
-            return "Mobile sync request timed out"
-        case .insecureManualRoute:
-            return "Manual host did not advertise a secure mobile sync route"
-        case .attachTicketExpired:
-            return "Mobile attach ticket expired"
-        case let .authorizationFailed(message):
-            return message
-        case let .rpcError(_, message):
-            return message
-        }
-    }
-}
-
-enum CmxAttachTicketInput {
-    static func decode(_ rawValue: String) throws -> CmxAttachTicket {
-        guard let url = URL(string: rawValue) else {
-            throw MobileSyncPairingPayloadError.invalidURL
-        }
-        if url.scheme == "cmux-ios", url.host == "pair" {
-            return try ticket(from: MobileSyncPairingPayload.decodeURL(url))
-        }
-        guard url.scheme == "cmux-ios",
-              url.host == "attach",
-              let components = URLComponents(url: url, resolvingAgainstBaseURL: false),
-              let encodedPayload = components.queryItems?.first(where: { $0.name == "payload" })?.value,
-              let data = base64URLDecode(encodedPayload) else {
-            throw MobileSyncPairingPayloadError.invalidURL
-        }
-        let decoder = JSONDecoder()
-        decoder.dateDecodingStrategy = .iso8601
-        let ticket = try decoder.decode(CmxAttachTicket.self, from: data)
-        try ticket.validate()
-        return ticket
-    }
-
-    private static func ticket(from payload: MobileSyncPairingPayload) throws -> CmxAttachTicket {
-        let route = try CmxAttachRoute(
-            id: payload.transport.rawValue,
-            kind: payload.transport,
-            endpoint: .hostPort(host: payload.host, port: payload.port)
-        )
-        let ticket = try CmxAttachTicket(
-            workspaceID: "workspace-main",
-            terminalID: nil,
-            macDeviceID: payload.macDeviceID,
-            macDisplayName: payload.macDisplayName,
-            routes: [route],
-            expiresAt: payload.expiresAt
-        )
-        try ticket.validate()
-        return ticket
-    }
-
-    private static func base64URLDecode(_ value: String) -> Data? {
-        var base64 = value
-            .replacingOccurrences(of: "-", with: "+")
-            .replacingOccurrences(of: "_", with: "/")
-        let padding = base64.count % 4
-        if padding > 0 {
-            base64.append(String(repeating: "=", count: 4 - padding))
-        }
-        return Data(base64Encoded: base64)
-    }
-}
-
-// All stored properties are immutable `let`s of Sendable types (the session is
-// an actor), so this is genuinely `Sendable` without opting out of checking.
-final class MobileCoreRPCClient: Sendable {
-    private let runtime: CMUXMobileRuntime
-    private let route: CmxAttachRoute
-    private let ticket: CmxAttachTicket
-    private let allowsStackAuthFallback: Bool
-    private let session: MobileCoreRPCSession
-
-    init(
-        runtime: CMUXMobileRuntime,
-        route: CmxAttachRoute,
-        ticket: CmxAttachTicket,
-        allowsStackAuthFallback: Bool = false
-    ) {
-        self.runtime = runtime
-        self.route = route
-        self.ticket = ticket
-        self.allowsStackAuthFallback = allowsStackAuthFallback
-        self.session = MobileCoreRPCSession(
-            makeTransport: { [route, runtime] in
-                try runtime.transportFactory.makeTransport(for: route)
-            }
-        )
-    }
-
-    /// Tear down the persistent transport (called when the client is
-    /// replaced or the user signs out).
-    func disconnect() async {
-        await session.tearDown(error: .connectionClosed)
-    }
-
-    /// Subscribe to server-pushed events. Returns a stream of envelopes
-    /// matching any of the requested topics. Cancel by terminating iteration.
-    func subscribe(to topics: Set<String>) async -> AsyncStream<MobileEventEnvelope> {
-        await session.addEventListener(topics: topics).stream
-    }
-
-    static func requestData(
-        method: String,
-        params: [String: Any] = [:],
-        id: String = UUID().uuidString
-    ) throws -> Data {
-        let request: [String: Any] = [
-            "id": id,
-            "method": method,
-            "params": params,
-        ]
-        return try JSONSerialization.data(withJSONObject: request)
-    }
-
-    func sendRequest(_ requestData: Data, timeoutNanoseconds: UInt64? = nil) async throws -> Data {
-        // Multiplexed over a persistent transport: each request gets a unique
-        // id, the session's reader task routes the response back here. No
-        // connect/close per RPC, no head-of-line blocking between calls.
-        let (id, augmented) = try Self.requestWithGuaranteedID(requestData)
-        let authenticated = try await requestDataWithAuth(augmented)
-        return try await Self.withRequestTimeout(
-            timeoutNanoseconds: timeoutNanoseconds ?? runtime.rpcRequestTimeoutNanoseconds
-        ) {
-            try await self.session.send(payload: authenticated, requestID: id)
-        }
-    }
-
-    private static func requestWithGuaranteedID(_ requestData: Data) throws -> (String, Data) {
-        guard var dict = try JSONSerialization.jsonObject(with: requestData) as? [String: Any] else {
-            throw MobileShellConnectionError.invalidResponse
-        }
-        let id: String
-        if let existing = dict["id"] as? String, !existing.isEmpty {
-            id = existing
-        } else {
-            id = UUID().uuidString
-            dict["id"] = id
-        }
-        let data = try JSONSerialization.data(withJSONObject: dict)
-        return (id, data)
-    }
-
-    private func requestDataWithAuth(_ requestData: Data) async throws -> Data {
-        guard var request = try JSONSerialization.jsonObject(with: requestData) as? [String: Any] else {
-            return requestData
-        }
-        let requestNeedsAuth = Self.requestRequiresAuth(request)
-        let requestIsCoveredByAttachTicket = !Self.requestNeedsStackAuthFallback(request, ticket: ticket)
-        var auth: [String: Any] = [:]
-        let attachToken = ticket.authToken?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let hasAttachToken = attachToken?.isEmpty == false
-        if let attachToken,
-           requestNeedsAuth,
-           hasAttachToken,
-           requestIsCoveredByAttachTicket {
-            if ticket.expiresAt > runtime.now() {
-                auth["attach_token"] = attachToken
-            } else if !allowsStackAuthFallback || !MobileShellRouteAuthPolicy.routeAllowsStackAuth(route) {
-                throw MobileShellConnectionError.attachTicketExpired
-            }
-        }
-        let shouldSendStackAuth = requestNeedsAuth && auth["attach_token"] == nil
-        if shouldSendStackAuth {
-            guard allowsStackAuthFallback,
-                  MobileShellRouteAuthPolicy.routeAllowsStackAuth(route) else {
-                throw MobileShellConnectionError.insecureManualRoute
-            }
-            do {
-                auth["stack_access_token"] = try await runtime.stackAccessTokenProvider()
-            } catch {
-                throw MobileShellConnectionError.authorizationFailed(
-                    L10n.string(
-                        "mobile.pairing.stackAuthTokenUnavailable",
-                        defaultValue: "Sign in on your computer with the same account, then try again."
-                    )
-                )
-            }
-        }
-        if !auth.isEmpty {
-            request["auth"] = auth
-        }
-        return try JSONSerialization.data(withJSONObject: request)
-    }
-
-    private static func requestNeedsStackAuthFallback(_ request: [String: Any], ticket: CmxAttachTicket) -> Bool {
-        guard requestRequiresAuth(request) else {
-            return false
-        }
-        let method = (request["method"] as? String)?.trimmingCharacters(in: .whitespacesAndNewlines)
-        let params = request["params"] as? [String: Any] ?? [:]
-        let workspaceSelection = stringParamSelection(params, keys: ["workspace_id"])
-        let terminalSelection = stringParamSelection(params, keys: ["surface_id", "terminal_id", "tab_id"])
-        if workspaceSelection.hasConflict ||
-            terminalSelection.hasConflict ||
-            containsIgnoredAliasParameters(params) {
-            return true
-        }
-
-        switch method {
-        case "mobile.workspace.list", "workspace.list":
-            return false
-        case "workspace.create":
-            return false
-        case "mobile.terminal.create", "terminal.create":
-            return false
-        case "mobile.terminal.input", "terminal.input",
-             "mobile.terminal.replay", "terminal.replay",
-             "mobile.terminal.viewport", "terminal.viewport":
-            return !ticketCoversTerminalRequest(
-                ticket: ticket,
-                workspaceSelection: workspaceSelection.value,
-                terminalSelection: terminalSelection.value
-            )
-        case "mobile.events.subscribe", "mobile.events.unsubscribe":
-            return false
-        default:
-            return true
-        }
-    }
-
-    private static func requestRequiresAuth(_ request: [String: Any]) -> Bool {
-        let method = (request["method"] as? String)?.trimmingCharacters(in: .whitespacesAndNewlines)
-        // Only the unauthenticated host probe is exempt. attach_ticket.create has no
-        // attach token yet (it mints the ticket), so requiring auth routes it through
-        // the Stack Auth account token: a ticket can only be created by a signed-in user.
-        return method != "mobile.host.status"
-    }
-
-    private static func ticketCoversTerminalRequest(
-        ticket: CmxAttachTicket,
-        workspaceSelection: String?,
-        terminalSelection: String?
-    ) -> Bool {
-        let ticketWorkspaceID = ticket.workspaceID.trimmingCharacters(in: .whitespacesAndNewlines)
-        // Empty workspaceID means the ticket is Mac-wide (general pairing).
-        // It covers any workspace/terminal on the paired Mac.
-        if ticketWorkspaceID.isEmpty {
-            return true
-        }
-        if let workspaceSelection, workspaceSelection != ticketWorkspaceID {
-            return false
-        }
-
-        if let ticketTerminalID = ticket.terminalID?.trimmingCharacters(in: .whitespacesAndNewlines),
-           !ticketTerminalID.isEmpty {
-            return terminalSelection == ticketTerminalID
-        }
-
-        return workspaceSelection == ticketWorkspaceID
-    }
-
-    private static func containsIgnoredAliasParameters(_ params: [String: Any]) -> Bool {
-        params["workspaceID"] != nil || params["terminalID"] != nil
-    }
-
-    private static func stringParamSelection(
-        _ params: [String: Any],
-        keys: [String]
-    ) -> StringParamSelection {
-        var selected: String?
-        for key in keys {
-            if let value = params[key] as? String {
-                let trimmed = value.trimmingCharacters(in: .whitespacesAndNewlines)
-                if !trimmed.isEmpty {
-                    if let selected, selected != trimmed {
-                        return StringParamSelection(value: selected, hasConflict: true)
-                    }
-                    selected = selected ?? trimmed
-                }
-            }
-        }
-        return StringParamSelection(value: selected, hasConflict: false)
-    }
-
-    private struct StringParamSelection {
-        let value: String?
-        let hasConflict: Bool
-    }
-
-    private static func withRequestTimeout<T: Sendable>(
-        timeoutNanoseconds: UInt64,
-        operation: @escaping @Sendable () async throws -> T
-    ) async throws -> T {
-        try Task.checkCancellation()
-        return try await withThrowingTaskGroup(of: T.self) { group in
-            group.addTask {
-                try await operation()
-            }
-            group.addTask {
-                try await Task.sleep(nanoseconds: timeoutNanoseconds)
-                throw MobileShellConnectionError.requestTimedOut
-            }
-            do {
-                guard let result = try await group.next() else {
-                    throw CancellationError()
-                }
-                group.cancelAll()
-                return result
-            } catch {
-                group.cancelAll()
-                throw error
-            }
-        }
-    }
-}
-
-#if DEBUG
-extension MobileCoreRPCClient {
-    static func debugWithRequestTimeout<T: Sendable>(
-        timeoutNanoseconds: UInt64,
-        operation: @escaping @Sendable () async throws -> T
-    ) async throws -> T {
-        try await withRequestTimeout(
-            timeoutNanoseconds: timeoutNanoseconds,
-            operation: operation
-        )
-    }
-}
-#endif
-
-extension CmxAttachEndpoint {
-    var logDescription: String {
-        switch self {
-        case let .hostPort(host, port):
-            return "\(host):\(port)"
-        case let .peer(id, relayHint, directAddrs, relayURL):
-            let addressSummary = directAddrs.isEmpty ? "no-direct-addrs" : "\(directAddrs.count)-direct-addrs"
-            return "peer:\(id):\(relayHint ?? relayURL ?? "no-relay"):\(addressSummary)"
-        case let .url(url):
-            return url
-        }
-    }
-}
-
-struct MobileSyncWorkspaceListResponse: Decodable, Sendable {
-    struct Workspace: Decodable, Sendable {
-        let id: String
-        let title: String
-        let currentDirectory: String?
-        let isSelected: Bool
-        let terminals: [Terminal]
-
-        private enum CodingKeys: String, CodingKey {
-            case id
-            case title
-            case currentDirectory = "current_directory"
-            case isSelected = "is_selected"
-            case terminals
-        }
-    }
-
-    struct Terminal: Decodable, Sendable {
-        let id: String
-        let title: String
-        let currentDirectory: String?
-        let isFocused: Bool
-        let isReady: Bool?
-
-        private enum CodingKeys: String, CodingKey {
-            case id
-            case title
-            case currentDirectory = "current_directory"
-            case isFocused = "is_focused"
-            case isReady = "is_ready"
-        }
-    }
-
-    let workspaces: [Workspace]
-    let createdWorkspaceID: String?
-    let createdTerminalID: String?
-
-    private enum CodingKeys: String, CodingKey {
-        case workspaces
-        case createdWorkspaceID = "created_workspace_id"
-        case createdTerminalID = "created_terminal_id"
-    }
-
-    static func decode(_ data: Data) throws -> MobileSyncWorkspaceListResponse {
-        try JSONDecoder().decode(Self.self, from: data)
-    }
-}
-
-extension MobileWorkspacePreview {
-    init(remote: MobileSyncWorkspaceListResponse.Workspace) {
-        self.init(
-            id: ID(rawValue: remote.id),
-            name: remote.title,
-            terminals: remote.terminals.map { terminal in
-                MobileTerminalPreview(remote: terminal)
-            }
-        )
-    }
-}
-
-extension MobileTerminalPreview {
-    init(remote: MobileSyncWorkspaceListResponse.Terminal) {
-        self.init(
-            id: ID(rawValue: remote.id),
-            name: remote.title,
-            isReady: remote.isReady ?? true,
-            isFocused: remote.isFocused
-        )
-    }
-}
-
-enum PreviewMobileHost {
-    static let hostName = "cmux-macbook"
-
-    static let workspaces: [MobileWorkspacePreview] = [
-        MobileWorkspacePreview(
-            id: "workspace-main",
-            name: "cmux",
-            terminals: [
-                MobileTerminalPreview(id: "terminal-build", name: "Build"),
-                MobileTerminalPreview(id: "terminal-agent", name: "Agent"),
-                MobileTerminalPreview(id: "terminal-tui", name: "TUI"),
-            ]
-        ),
-        MobileWorkspacePreview(
-            id: "workspace-docs",
-            name: "Docs",
-            terminals: [
-                MobileTerminalPreview(id: "terminal-notes", name: "Notes"),
-            ]
-        ),
-    ]
-}
-
-// MARK: - MobileCoreRPCSession
-
-/// One server-pushed event delivered over the persistent transport.
-public struct MobileEventEnvelope: Sendable {
-    public let topic: String
-    public let payloadJSON: Data?
-    public let streamID: String?
-}
-
-/// Owns a single persistent transport for a `MobileCoreRPCClient`, multiplexes
-/// requests by id, and dispatches server-pushed events to registered listeners.
-/// No polling: the reader task runs continuously, parking on `transport.receive()`
-/// until the kernel delivers bytes. There is no `Task.sleep` or `asyncAfter`
-/// anywhere in this class; the only Task.sleep elsewhere in the file is the
-/// race-deadline in `withRequestTimeout`.
-private actor MobileCoreRPCSession {
-    typealias TransportFactory = @Sendable () throws -> any CmxByteTransport
-    typealias PendingContinuation = CheckedContinuation<Result<Data, MobileShellConnectionError>, Never>
-
-    struct EventSubscription {
-        let id: UUID
-        let stream: AsyncStream<MobileEventEnvelope>
-    }
-
-    private struct EventListener {
-        let topics: Set<String>
-        let continuation: AsyncStream<MobileEventEnvelope>.Continuation
-    }
-
-    private struct PendingWrite: Sendable {
-        let requestID: String
-        let frame: Data
-    }
-
-    private let makeTransport: TransportFactory
-    private var transport: (any CmxByteTransport)?
-    private var connectionTask: (id: UUID, task: Task<any CmxByteTransport, Error>)?
-    private var installedConnectionID: UUID?
-    private var readerTask: Task<Void, Never>?
-    private var pending: [String: PendingContinuation] = [:]
-    private var queuedRequestIDs: Set<String> = []
-    private var cancelledQueuedRequestIDs: Set<String> = []
-    private var listeners: [UUID: EventListener] = [:]
-    private var isTearingDown: Bool = false
-    /// Pending writes drained by `writerTask`. Serializes `transport.send` so
-    /// two concurrent `send(payload:requestID:)` callers never trip
-    /// `CmxNetworkByteTransport.sendAlreadyInProgress`. AsyncStream backed so
-    /// the writer parks on `await` instead of polling.
-    private var writeQueue: AsyncStream<PendingWrite>.Continuation?
-    private var writerTask: Task<Void, Never>?
-
-    init(makeTransport: @escaping TransportFactory) {
-        self.makeTransport = makeTransport
-    }
-
-    deinit {
-        connectionTask?.task.cancel()
-        readerTask?.cancel()
-        writerTask?.cancel()
-        writeQueue?.finish()
-    }
-
-    func send(payload: Data, requestID: String) async throws -> Data {
-        _ = try await ensureConnected()
-        let frame = try MobileSyncFrameCodec.encodeFrame(payload)
-
-        let result: Result<Data, MobileShellConnectionError> = await withTaskCancellationHandler {
-            await withCheckedContinuation { continuation in
-                // Register BEFORE handing the frame to the writer so a fast
-                // response can't race past us. Writer pulls frames serially
-                // from `writeQueue`, so concurrent senders never overlap a
-                // `transport.send()` call.
-                pending[requestID] = continuation
-                guard let queue = writeQueue else {
-                    pending.removeValue(forKey: requestID)
-                    continuation.resume(returning: .failure(.connectionClosed))
-                    return
-                }
-                queuedRequestIDs.insert(requestID)
-                _ = queue.yield(PendingWrite(requestID: requestID, frame: frame))
-            }
-        } onCancel: {
-            Task {
-                await self.cancelPendingRequest(requestID: requestID)
-            }
-        }
-        switch result {
-        case .success(let data):
-            return data
-        case .failure(let error):
-            throw error
-        }
-    }
-
-    func addEventListener(topics: Set<String>) -> EventSubscription {
-        let id = UUID()
-        var continuation: AsyncStream<MobileEventEnvelope>.Continuation!
-        let stream = AsyncStream<MobileEventEnvelope>(bufferingPolicy: .bufferingNewest(256)) { cont in
-            continuation = cont
-        }
-        listeners[id] = EventListener(topics: topics, continuation: continuation)
-        continuation.onTermination = { [weak self] _ in
-            guard let self else { return }
-            Task { await self.removeListener(id: id) }
-        }
-        return EventSubscription(id: id, stream: stream)
-    }
-
-    func removeListener(id: UUID) {
-        listeners.removeValue(forKey: id)
-    }
-
-    func tearDown(error: MobileShellConnectionError) async {
-        guard !isTearingDown else { return }
-        isTearingDown = true
-        let pendingSnapshot = pending
-        pending.removeAll()
-        queuedRequestIDs.removeAll()
-        cancelledQueuedRequestIDs.removeAll()
-        for (_, cont) in pendingSnapshot {
-            cont.resume(returning: .failure(error))
-        }
-        let listenerSnapshot = listeners
-        listeners.removeAll()
-        for (_, listener) in listenerSnapshot {
-            listener.continuation.finish()
-        }
-        // Stop the writer loop before closing the transport so we don't try to
-        // write into a half-closed socket and never trigger
-        // sendAlreadyInProgress on a torn-down state.
-        writeQueue?.finish()
-        writeQueue = nil
-        writerTask?.cancel()
-        writerTask = nil
-        connectionTask?.task.cancel()
-        connectionTask = nil
-        installedConnectionID = nil
-        if let transport {
-            await transport.close()
-        }
-        transport = nil
-        readerTask?.cancel()
-        readerTask = nil
-        isTearingDown = false
-    }
-
-    // MARK: - private
-
-    private func ensureConnected() async throws -> any CmxByteTransport {
-        if let transport { return transport }
-
-        let connectionID: UUID
-        let task: Task<any CmxByteTransport, Error>
-        if let existing = connectionTask {
-            connectionID = existing.id
-            task = existing.task
-        } else {
-            let candidate = try makeTransport()
-            connectionID = UUID()
-            task = Task {
-                try await candidate.connect()
-                return candidate
-            }
-            connectionTask = (id: connectionID, task: task)
-        }
-
-        let candidate: any CmxByteTransport
-        do {
-            candidate = try await task.value
-        } catch {
-            if connectionTask?.id == connectionID {
-                connectionTask = nil
-            }
-            throw error
-        }
-
-        if let transport {
-            if installedConnectionID != connectionID {
-                await candidate.close()
-            }
-            return transport
-        }
-
-        guard connectionTask?.id == connectionID else {
-            await candidate.close()
-            throw MobileShellConnectionError.connectionClosed
-        }
-
-        connectionTask = nil
-        installedConnectionID = connectionID
-        transport = candidate
-        // Reader: dispatches inbound frames by id (response) or topic (event).
-        readerTask = Task { [weak self] in
-            await self?.readLoop(transport: candidate)
-        }
-        // Writer: drains queued frames one at a time so concurrent send()
-        // callers don't trigger CmxNetworkByteTransport.sendAlreadyInProgress.
-        // Failures tear the whole session down which fails every pending
-        // continuation.
-        let (stream, continuation) = AsyncStream<PendingWrite>.makeStream(bufferingPolicy: .unbounded)
-        writeQueue = continuation
-        writerTask = Task { [weak self] in
-            await self?.writeLoop(transport: candidate, frames: stream)
-        }
-        return candidate
-    }
-
-    private func writeLoop(transport: any CmxByteTransport, frames: AsyncStream<PendingWrite>) async {
-        for await write in frames {
-            if Task.isCancelled { return }
-            guard shouldSendQueuedWrite(requestID: write.requestID) else {
-                continue
-            }
-            do {
-                try await transport.send(write.frame)
-            } catch {
-                await tearDown(error: .connectionClosed)
-                return
-            }
-        }
-    }
-
-    private func readLoop(transport: any CmxByteTransport) async {
-        var buffer = Data()
-        while !Task.isCancelled {
-            let chunk: Data?
-            do {
-                chunk = try await transport.receive()
-            } catch {
-                await tearDown(error: .connectionClosed)
-                return
-            }
-            guard let chunk, !chunk.isEmpty else {
-                if chunk == nil {
-                    await tearDown(error: .connectionClosed)
-                    return
-                }
-                continue
-            }
-            buffer.append(chunk)
-            let frames: [Data]
-            do {
-                frames = try MobileSyncFrameCodec.decodeFrames(from: &buffer)
-            } catch {
-                await tearDown(error: .invalidResponse)
-                return
-            }
-            for frame in frames {
-                dispatch(frame: frame)
-            }
-        }
-    }
-
-    private func dispatch(frame: Data) {
-        let parsed = try? JSONSerialization.jsonObject(with: frame) as? [String: Any]
-        guard let envelope = parsed else { return }
-        if (envelope["kind"] as? String) == "event" {
-            guard let topic = envelope["topic"] as? String else { return }
-            let payloadData: Data?
-            if let payload = envelope["payload"] {
-                payloadData = try? JSONSerialization.data(withJSONObject: payload)
-            } else {
-                payloadData = nil
-            }
-            let streamID = envelope["stream_id"] as? String
-            let event = MobileEventEnvelope(topic: topic, payloadJSON: payloadData, streamID: streamID)
-            for (_, listener) in listeners where listener.topics.contains(topic) {
-                listener.continuation.yield(event)
-            }
-            return
-        }
-        guard let id = envelope["id"] as? String else { return }
-        guard let cont = pending.removeValue(forKey: id) else { return }
-        if (envelope["ok"] as? Bool) == true {
-            let result = envelope["result"] ?? [:]
-            if let data = try? JSONSerialization.data(withJSONObject: result) {
-                cont.resume(returning: .success(data))
-            } else {
-                cont.resume(returning: .failure(.invalidResponse))
-            }
-            return
-        }
-        let errorPayload = envelope["error"] as? [String: Any]
-        let message = (errorPayload?["message"] as? String) ?? "RPC error"
-        let code = errorPayload?["code"] as? String
-        if code == "unauthorized" {
-            cont.resume(returning: .failure(.authorizationFailed(message)))
-        } else {
-            cont.resume(returning: .failure(.rpcError(code, message)))
-        }
-    }
-
-    private func failPending(requestID: String, error: MobileShellConnectionError) {
-        guard let cont = pending.removeValue(forKey: requestID) else { return }
-        cont.resume(returning: .failure(error))
-    }
-
-    private func cancelPendingRequest(requestID: String) {
-        guard let cont = pending.removeValue(forKey: requestID) else { return }
-        if queuedRequestIDs.remove(requestID) != nil {
-            cancelledQueuedRequestIDs.insert(requestID)
-        }
-        cont.resume(returning: .failure(.requestTimedOut))
-    }
-
-    private func shouldSendQueuedWrite(requestID: String) -> Bool {
-        let wasQueued = queuedRequestIDs.remove(requestID) != nil
-        if cancelledQueuedRequestIDs.remove(requestID) != nil {
-            return false
-        }
-        return wasQueued && pending[requestID] != nil
-    }
-}
