+++ ios/cmuxPackage/Tests/cmuxFeatureTests/cmuxFeatureTests.swift
@@ -1,5 +1,7 @@
 import CMUXMobileCore
 @testable import CmuxMobileAuth
+import CmuxMobilePairedMac
+import CmuxMobileRPC
 import CmuxMobileShellModel
 import CmuxMobileWorkspace
 import Foundation
@@ -111,116 +113,6 @@ import UIKit
     )
 }
 
-@Test func rpcRequestTimeoutCancelsOperationWhenCallerIsCancelled() async throws {
-    let started = AsyncFlag()
-    let cancelled = AsyncFlag()
-    let task = Task {
-        try await MobileCoreRPCClient.debugWithRequestTimeout(
-            timeoutNanoseconds: 60 * 1_000_000_000
-        ) {
-            await started.set()
-            do {
-                try await Task.sleep(nanoseconds: 60 * 1_000_000_000)
-                return "completed"
-            } catch {
-                await cancelled.set()
-                throw error
-            }
-        }
-    }
-
-    for _ in 0..<100 {
-        if await started.isSet() {
-            break
-        }
-        try await Task.sleep(nanoseconds: 1_000_000)
-    }
-    #expect(await started.isSet())
-
-    task.cancel()
-
-    do {
-        _ = try await task.value
-        Issue.record("Expected cancelled RPC timeout wrapper to throw")
-    } catch is CancellationError {
-    } catch {
-        Issue.record("Expected CancellationError, got \(error)")
-    }
-
-    for _ in 0..<100 {
-        if await cancelled.isSet() {
-            break
-        }
-        try await Task.sleep(nanoseconds: 1_000_000)
-    }
-    #expect(await cancelled.isSet())
-}
-
-@Test func cancelledQueuedRPCIsNotWrittenAfterEarlierSendCompletes() async throws {
-    let transport = QueuedCancellationProbeTransport()
-    let route = try hostPortRoute(kind: .debugLoopback, host: "127.0.0.1", port: 59123)
-    let runtime = testRuntime(
-        transportFactory: QueuedCancellationProbeTransportFactory(transport: transport),
-        rpcRequestTimeoutNanoseconds: 60 * 1_000_000_000
-    )
-    let ticket = try CmxAttachTicket(
-        workspaceID: "workspace-main",
-        terminalID: "terminal-main",
-        macDeviceID: "test-mac",
-        macDisplayName: "Test Mac",
-        routes: [route],
-        expiresAt: Date().addingTimeInterval(60),
-        authToken: "ticket-secret"
-    )
-    let client = MobileCoreRPCClient(runtime: runtime, route: route, ticket: ticket)
-    let firstRequest = try MobileCoreRPCClient.requestData(
-        method: "terminal.input",
-        params: [
-            "workspace_id": "workspace-main",
-            "terminal_id": "terminal-main",
-            "text": "first",
-        ],
-        id: "first-input"
-    )
-    let queuedRequest = try MobileCoreRPCClient.requestData(
-        method: "workspace.create",
-        params: ["title": "queued-workspace"],
-        id: "queued-create"
-    )
-
-    let firstTask = Task {
-        try await client.sendRequest(firstRequest)
-    }
-    let firstSent = try await transport.waitForSentRequestCount(1)
-    #expect(firstSent.map(\.method) == ["terminal.input"])
-
-    let queuedTask = Task {
-        try await client.sendRequest(queuedRequest)
-    }
-    for _ in 0..<100 {
-        await Task.yield()
-    }
-    queuedTask.cancel()
-    do {
-        _ = try await queuedTask.value
-        Issue.record("Expected queued RPC cancellation to throw")
-    } catch {
-    }
-
-    await transport.releaseFirstSend()
-    for _ in 0..<100 {
-        if try await transport.sentRequests().count > 1 {
-            break
-        }
-        try await Task.sleep(nanoseconds: 1_000_000)
-    }
-
-    let sent = try await transport.sentRequests()
-    #expect(sent.map(\.method) == ["terminal.input"])
-    firstTask.cancel()
-    _ = try? await firstTask.value
-}
-
 @Test func mobileRuntimeDefaultsToThirtySecondRPCTimeout() {
     let runtime = CMUXMobileRuntime(
         supportedRouteKinds: [.debugLoopback],
@@ -230,38 +122,6 @@ import UIKit
 
     #expect(runtime.rpcRequestTimeoutNanoseconds == 30 * 1_000_000_000)
     #expect(runtime.pairingRequestTimeoutNanoseconds == 8 * 1_000_000_000)
-}
-
-@Test func pairedMacStorePersistsActiveMacsScopedByStackUser() async throws {
-    let directory = FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString, isDirectory: true)
-    try FileManager.default.createDirectory(at: directory, withIntermediateDirectories: true)
-    defer { try? FileManager.default.removeItem(at: directory) }
-    let store = try MobilePairedMacStore(databaseURL: directory.appendingPathComponent("paired-macs.sqlite3"))
-    let userARoute = try hostPortRoute(kind: .tailscale, host: "work-a.tailnet.ts.net", port: CmxMobileDefaults.defaultHostPort)
-    let userBRoute = try hostPortRoute(kind: .tailscale, host: "work-b.tailnet.ts.net", port: CmxMobileDefaults.defaultHostPort)
-    let now = Date(timeIntervalSince1970: 1_000)
-
-    try await store.upsert(
-        macDeviceID: "mac-a",
-        displayName: "Work A",
-        routes: [userARoute],
-        markActive: true,
-        stackUserID: "user-a",
-        now: now
-    )
-    try await store.upsert(
-        macDeviceID: "mac-b",
-        displayName: "Work B",
-        routes: [userBRoute],
-        markActive: true,
-        stackUserID: "user-b",
-        now: now.addingTimeInterval(10)
-    )
-
-    #expect(try await store.activeMac(stackUserID: "user-a")?.macDeviceID == "mac-a")
-    #expect(try await store.activeMac(stackUserID: "user-b")?.macDeviceID == "mac-b")
-    #expect(try await store.activeMac(stackUserID: "missing") == nil)
-    #expect(try await store.loadAll(stackUserID: "user-a").map(\.routes).first == [userARoute])
 }
 
 @MainActor
@@ -2365,14 +2225,6 @@ private struct ScriptedTransportFactory: CmxByteTransportFactory {
     }
 }
 
-private struct QueuedCancellationProbeTransportFactory: CmxByteTransportFactory {
-    let transport: QueuedCancellationProbeTransport
-
-    func makeTransport(for route: CmxAttachRoute) throws -> any CmxByteTransport {
-        transport
-    }
-}
-
 private struct FailingRouteTransportFactory: CmxByteTransportFactory {
     let failingRouteID: String
     let responses: ScriptedTransportResponses
@@ -2888,18 +2740,6 @@ private struct RecordedRPCRequest: Sendable {
     var stackAccessToken: String?
 }
 
-private actor AsyncFlag {
-    private var value = false
-
-    func set() {
-        value = true
-    }
-
-    func isSet() -> Bool {
-        value
-    }
-}
-
 private func recordedRPCRequest(from payload: Data) throws -> RecordedRPCRequest {
     let request = try #require(JSONSerialization.jsonObject(with: payload) as? [String: Any])
     let params = request["params"] as? [String: Any] ?? [:]
@@ -3008,68 +2848,6 @@ private actor ScriptedTransport: CmxByteTransport {
         for waiter in waiters {
             waiter.resume(returning: nil)
         }
-    }
-}
-
-private actor QueuedCancellationProbeTransport: CmxByteTransport {
-    private var sentPayloads: [Data] = []
-    private var receiveWaiters: [CheckedContinuation<Data?, Never>] = []
-    private var firstSendRelease: CheckedContinuation<Void, Never>?
-    private var shouldBlockFirstSend = true
-    private var isClosed = false
-
-    func connect() async throws {}
-
-    func receive() async throws -> Data? {
-        if isClosed {
-            return nil
-        }
-        return await withCheckedContinuation { continuation in
-            receiveWaiters.append(continuation)
-        }
-    }
-
-    func send(_ data: Data) async throws {
-        var buffer = data
-        let payloads = try MobileSyncFrameCodec.decodeFrames(from: &buffer)
-        sentPayloads.append(contentsOf: payloads)
-        if shouldBlockFirstSend {
-            shouldBlockFirstSend = false
-            await withCheckedContinuation { continuation in
-                firstSendRelease = continuation
-            }
-        }
-    }
-
-    func close() async {
-        isClosed = true
-        let waiters = receiveWaiters
-        receiveWaiters = []
-        for waiter in waiters {
-            waiter.resume(returning: nil)
-        }
-        releaseFirstSend()
-    }
-
-    func releaseFirstSend() {
-        firstSendRelease?.resume()
-        firstSendRelease = nil
-    }
-
-    func sentRequests() throws -> [RecordedRPCRequest] {
-        try sentPayloads.map(recordedRPCRequest(from:))
-    }
-
-    func waitForSentRequestCount(_ count: Int) async throws -> [RecordedRPCRequest] {
-        var requests: [RecordedRPCRequest] = []
-        for _ in 0..<200 {
-            requests = try sentRequests()
-            if requests.count >= count {
-                return requests
-            }
-            try await Task.sleep(nanoseconds: 1_000_000)
-        }
-        return requests
     }
 }
